Feat : Implement a feature making a comprehensive data file 다수의 파일을 하나로 통합해주는 기능 구현
